Moved around files and again cleaned up lint issues
