+++ src/components/NavigationBar/InternalNavBar.jsx
@@ -1,7 +1,7 @@
 import { Link } from 'react-router-dom';
-import '../NavigationBar.css';
-import logo from '../../../images/logo2.png';
-import NavigationBarItem from '../NavigationBarDef';
+import './NavigationBar.css';
+import logo from '../../images/logo2.png';
+import NavigationBarItem from './NavigationBarDef';
 
 function InternalNavigationBar() {
   // **** @todo Fix the linking of the nav bar items ***
