+++ src/components/common/DisplayCheckboxGroup.jsx
@@ -1,3 +1,4 @@
+/* eslint-disable jsx-a11y/label-has-associated-control */
 import React from 'react';
 
 /**
