+++ src/pages/Listing/ViewListingPage.jsx
@@ -12,6 +12,7 @@ function ViewListingPage() {
 
   // TODO for now dont do this check for dev purposes
   // if the current user is not logged in
+  // eslint-disable-next-line no-constant-condition
   if (!currentUser && false) {
     navigate('../..', { relative: 'path' });
   }
@@ -48,11 +49,7 @@ function ViewListingPage() {
     }
   }, [listingId]);
 
-  return (
-    <>
-      <div>{listingComponent}</div>
-    </>
-  );
+  return <div>{listingComponent}</div>;
 }
 
 export default ViewListingPage;
