+++ src/pages/LoginPage.jsx
@@ -12,8 +12,7 @@ function LoginPage() {
       <AuthProvider>
         <Container
           className="d-flex align-items-center justify-content-center"
-          style={{ minHeight: '80vh' }}
-        >
+          style={{ minHeight: '80vh' }}>
           <div className="w-100" style={{ maxWidth: '400px' }}>
             <Signin />
           </div>
